Move LoginViewTest to core app, create login function for board/test_views :baby: #1
