+++ zendesk_tickets_machine/zendesk_tickets_machine/settings/base.py
@@ -40,6 +40,7 @@ INSTALLED_APPS = [
     'agents',
     'agent_groups',
     'boards',
+    'core',
     'requesters',
     'tickets',
 ]
